get reprocessed tree dataset working

+++ src/main/java/io/github/terra121/util/TerraMath.java
@@ -4,7 +4,7 @@ import lombok.experimental.UtilityClass;
 
 @UtilityClass
 public class TerraMath {
-    public double lerp(int a, int b, int t) {
+    public double lerp(double a, double b, double t) {
         return a + (b - a) * t;
     }
 }
